requirements.txt
torch>=2.0.0
numpy>=1.21.0
jieba>=0.42.1
nltk>=3.8.1
sentencepiece>=0.1.99
pyyaml>=6.0
tqdm>=4.65.0
sacrebleu>=2.3.1
tensorboard>=2.12.0
matplotlib>=3.5.0

